+++ src/core/blob.h
@@ -14,8 +14,8 @@ namespace Lux
 			Blob();
 
 			void create(void* data, int size) { m_data = data; m_size = size; m_pos = 0; }
-			virtual void write(const void* data, size_t size);
-			virtual bool read(void* data, size_t size);
+			virtual void write(const void* data, int32_t size);
+			virtual bool read(void* data, int32_t size);
 			const uint8_t* getBuffer() const { return &m_buffer[0]; }
 			int getBufferSize() const { return m_size; }
 			void flush() { m_size = 0; }
